New ot model version (#26)
* new mzn model working

* cleanup

* buses

* PR suggestions applied

+++ src/main/java/app/backend/document/CarFlow.java
@@ -33,7 +33,7 @@ public class CarFlow {
         this.id = id;
     }
 
-    public double getCarFlow() {
+    public int getCarFlow() {
         return carFlow;
     }
